Move encryption tests to arrow_reader tests
